Make DomainObjectCollectionFactory only for 5.1+ - CollectionCallbackActionDecorator is new in Gradle 5.1+
