+++ src/test/groovy/com/netflix/gradle/plugins/daemon/DaemonExtensionSpec.groovy
@@ -16,8 +16,7 @@
 
 package com.netflix.gradle.plugins.daemon
 
-import com.netflix.gradle.plugins.utils.BackwardsCompatibleDomainObjectCollectionFactory
-import nebula.test.IntegrationSpec
+import com.netflix.gradle.plugins.utils.DomainObjectCollectionFactory
 import nebula.test.ProjectSpec
 import org.gradle.api.internal.CollectionCallbackActionDecorator
 
@@ -26,7 +25,7 @@ class DaemonExtensionSpec extends ProjectSpec {
     def 'configures on add'() {
         given:
         def mockCollectionCallbackActionDecorator = Mock(CollectionCallbackActionDecorator)
-        BackwardsCompatibleDomainObjectCollectionFactory factory = new BackwardsCompatibleDomainObjectCollectionFactory<>(project.gradle.gradleVersion, mockCollectionCallbackActionDecorator)
+        DomainObjectCollectionFactory factory = new DomainObjectCollectionFactory<>(mockCollectionCallbackActionDecorator)
         def definitionList = factory.create(DaemonDefinition)
         DaemonExtension extension = new DaemonExtension(factory.create(DaemonDefinition, definitionList))
